Choar: first code cleanup
