+++ src/client/lib/txt-to-mdx/index.ts
@@ -1,12 +1,18 @@
-// src/client/lib/txt-to-mdx/index.ts
+// src/client/lib/txt-to-mdx/enhanced-converter.ts
 import path from 'path';
 import fs from 'fs/promises';
 import { syncWithVideo } from './sync-with-video';
 import { createDir } from '@/client/lib/utils/fs';
 import { ScraperFactory } from './scrapers';
 import { logger } from './scrapers/logger';
-import { AIServiceFactory } from '@/server/services/ai/ai-service.factory';
 import { AIServiceType } from '@/server/services/ai/types';
+import { JobProgressReporter } from '@/server/queue/job-progress-utils';
+import {
+	JobError,
+	JobErrorType,
+	processError,
+} from '@/server/queue/job-error-utils';
+import { AIServiceFactory } from '@/server/services/ai/ai-service.factory';
 
 /**
  * Result interface for the conversion process
@@ -18,21 +24,37 @@ export interface ConversionResult {
 }
 
 /**
- * Main converter class for transforming YouTube transcripts to MDX format
+ * Configuration options for the converter
  */
-export class TxtToMdxConverter {
+export interface ConverterOptions {
+	aiServiceType?: AIServiceType;
+	dataPath?: string;
+	tempDir?: string;
+	progressReporter?: JobProgressReporter;
+}
+
+/**
+ * Enhanced converter class for transforming YouTube transcripts to MDX format
+ * with progress reporting and improved error handling
+ */
+export class EnhancedTxtToMdxConverter {
 	private aiService;
 	private dataPath: string;
 	private tempDir: string;
+	private progressReporter?: JobProgressReporter;
 
-	constructor(
-		dataPath: string = path.join(process.cwd(), 'public/data'),
-		tempDir: string = path.join(process.cwd(), 'temp'),
-		aiServiceType?: AIServiceType
-	) {
-		this.aiService = AIServiceFactory.getService(aiServiceType);
-		this.dataPath = dataPath;
-		this.tempDir = tempDir;
+	constructor(options: ConverterOptions = {}) {
+		this.aiService = AIServiceFactory.getService(options.aiServiceType);
+		this.dataPath =
+			options.dataPath || path.join(process.cwd(), 'public', 'data');
+		this.tempDir = options.tempDir || path.join(process.cwd(), 'temp');
+		this.progressReporter = options.progressReporter;
+
+		logger.info('Enhanced TxtToMdxConverter initialized', {
+			aiService: options.aiServiceType || 'default',
+			dataPath: this.dataPath,
+			tempDir: this.tempDir,
+		});
 	}
 
 	/**
@@ -50,25 +72,46 @@ export class TxtToMdxConverter {
 		let srtPath: string | undefined;
 
 		try {
+			// Report initialization
+			await this.progress('INITIALIZED');
+
 			await this.validateInputs(url, playlistId);
+			logger.info('Input validation successful', { url, playlistId });
 
 			// Set up directories
-			const { topicPath: playlistPath } = await this.setupDirectories(playlistId);
+			const { topicPath: playlistPath } = await this.setupDirectories(
+				playlistId
+			);
+			logger.info('Directories set up', { playlistPath });
+
+			// Report downloading stage
+			await this.progress('DOWNLOADING');
 
 			// Download and extract content
 			const { videoId, title, files } = await this.downloadContent(url);
 			txtPath = files.txt;
 			srtPath = files.srt;
 
-			const txtContent = await fs.readFile(txtPath, 'utf-8');
+			await this.progress('TRANSCRIPT_EXTRACTED');
+			logger.info('Transcript extraction complete', { videoId, title });
 
-			// Process the content
+			const txtContent = await fs.readFile(txtPath, 'utf-8');
+			logger.info('Text content read', { size: txtContent.length });
+
+			// Process the content with AI
+			await this.progress('AI_PROCESSING_STARTED');
 			const processedContent = await this.processWithAI(
 				txtContent,
 				title
 			);
 
-			// Create and process MDX
+			await this.progress('AI_PROCESSING_COMPLETED');
+			logger.info('AI processing complete', {
+				processedSize: processedContent.length,
+			});
+
+			// Synchronize with video timestamps
+			await this.progress('SYNCHRONIZING');
 			const { finalPath } = await this.createMDX(
 				processedContent,
 				videoId,
@@ -76,20 +119,28 @@ export class TxtToMdxConverter {
 				srtPath
 			);
 
+			await this.progress('SAVING_TO_DATABASE');
 			logger.info('Conversion completed successfully', {
 				videoId,
 				title,
 				finalPath,
 			});
 
+			// Report completion
+			await this.progress('COMPLETED');
 			return {
 				mdxPath: finalPath,
 				videoId,
 				title,
 			};
 		} catch (error) {
-			logger.error('Error in conversion process:', error);
-			throw this.handleError(error);
+			const jobError = processError(error);
+			logger.error('Error in conversion process:', jobError);
+
+			// Report failure
+			await this.progressReporter?.reportFailure(jobError);
+
+			throw jobError;
 		} finally {
 			// Cleanup temporary files
 			await this.cleanup(
@@ -103,13 +154,19 @@ export class TxtToMdxConverter {
 	 */
 	private async validateInputs(url: string, topicId: string): Promise<void> {
 		if (!url || !topicId) {
-			throw new Error('URL and topicId are required');
+			throw new JobError(
+				'URL and topicId are required',
+				JobErrorType.INVALID_URL
+			);
 		}
 
 		try {
 			new URL(url);
 		} catch {
-			throw new Error('Invalid URL provided');
+			throw new JobError(
+				'Invalid URL provided',
+				JobErrorType.INVALID_URL
+			);
 		}
 	}
 
@@ -117,30 +174,46 @@ export class TxtToMdxConverter {
 	 * Sets up necessary directories
 	 */
 	private async setupDirectories(topicId: string) {
-		const topicPath = path.join(this.dataPath, topicId);
-		await createDir(topicPath);
-		await createDir(this.tempDir);
+		try {
+			const topicPath = path.join(this.dataPath, topicId);
+			await createDir(topicPath);
+			await createDir(this.tempDir);
 
-		return { topicPath };
+			return { topicPath };
+		} catch (error) {
+			throw new JobError(
+				'Failed to create directories',
+				JobErrorType.FILESYSTEM,
+				{ cause: error, details: { topicId } }
+			);
+		}
 	}
 
 	/**
 	 * Downloads and extracts content using appropriate scraper
 	 */
 	private async downloadContent(url: string) {
-		const scraper = ScraperFactory.getScraper(url);
+		try {
+			const scraper = ScraperFactory.getScraper(url);
 
-		logger.info('Starting transcript download...');
-		const result = await scraper.scrape(url, this.tempDir);
+			logger.info('Starting transcript download...');
+			const result = await scraper.scrape(url, this.tempDir);
 
-		logger.info('Transcript download complete', {
-			videoId: result.videoId,
-			title: result.title,
-		});
+			logger.info('Transcript download complete', {
+				videoId: result.videoId,
+				title: result.title,
+			});
 
-		await this.verifyDownloadedFiles(result.files);
+			await this.verifyDownloadedFiles(result.files);
 
-		return result;
+			return result;
+		} catch (error) {
+			throw new JobError(
+				'Failed to download transcript',
+				JobErrorType.TRANSCRIPT_DOWNLOAD,
+				{ cause: error, details: { url } }
+			);
+		}
 	}
 
 	/**
@@ -152,8 +225,10 @@ export class TxtToMdxConverter {
 			await fs.access(files.srt);
 			logger.info('Verified files exist', files);
 		} catch (error) {
-			throw new Error(
-				`Downloaded files not found: ${(error as Error).message}`
+			throw new JobError(
+				'Downloaded files not found',
+				JobErrorType.FILESYSTEM,
+				{ cause: error, details: files }
 			);
 		}
 	}
@@ -174,6 +249,10 @@ export class TxtToMdxConverter {
 			processedContent = processedContent
 				.replace(/{/g, '(')
 				.replace(/}/g, ')');
+
+			// Report halfway progress during AI processing
+			await this.progress('AI_PROCESSING_HALFWAY');
+
 			return processedContent;
 		} catch (error) {
 			logger.error('AI service error:', error);
@@ -181,10 +260,19 @@ export class TxtToMdxConverter {
 				logger.warn(
 					'Primary AI service quota exceeded, attempting fallback...'
 				);
+
+				// Try fallback service
 				this.aiService = AIServiceFactory.getService(); // Get fallback service
 				return await this.aiService.processContent(txtContent, title);
 			}
-			throw error;
+
+			throw new JobError(
+				'AI processing failed',
+				error instanceof Error && error.message === 'QUOTA_EXCEEDED'
+					? JobErrorType.AI_QUOTA_EXCEEDED
+					: JobErrorType.AI_SERVICE,
+				{ cause: error, retry: false }
+			);
 		}
 	}
 
@@ -197,17 +285,25 @@ export class TxtToMdxConverter {
 		playlistPath: string,
 		srtPath: string
 	) {
-		// Create temporary MDX
-		const tempMdxPath = path.join(this.tempDir, `${videoId}_temp.mdx`);
-		await fs.writeFile(tempMdxPath, content);
-		logger.info('Created temp MDX file', { tempMdxPath });
+		try {
+			// Create temporary MDX
+			const tempMdxPath = path.join(this.tempDir, `${videoId}_temp.mdx`);
+			await fs.writeFile(tempMdxPath, content);
+			logger.info('Created temp MDX file', { tempMdxPath });
 
-		// Create final MDX with timestamps
-		const finalPath = path.join(playlistPath, `${videoId}.mdx`);
-		await syncWithVideo(tempMdxPath, srtPath, finalPath);
-		logger.info('Created final MDX with timestamps', { finalPath });
+			// Create final MDX with timestamps
+			const finalPath = path.join(playlistPath, `${videoId}.mdx`);
+			await syncWithVideo(tempMdxPath, srtPath, finalPath);
+			logger.info('Created final MDX with timestamps', { finalPath });
 
-		return { tempMdxPath, finalPath };
+			return { tempMdxPath, finalPath };
+		} catch (error) {
+			throw new JobError(
+				'Failed to create MDX file',
+				JobErrorType.VIDEO_SYNC,
+				{ cause: error, details: { videoId } }
+			);
+		}
 	}
 
 	/**
@@ -227,12 +323,13 @@ export class TxtToMdxConverter {
 	}
 
 	/**
-	 * Handles and transforms errors
+	 * Update progress using the progress reporter
 	 */
-	private handleError(error: unknown): Error {
-		if (error instanceof Error) {
-			return new Error(`Conversion failed: ${error.message}`);
+	private async progress(
+		stage: keyof typeof import('@/server/queue/job-progress-utils').PROGRESS_STAGES
+	): Promise<void> {
+		if (this.progressReporter) {
+			await this.progressReporter.reportStage(stage);
 		}
-		return new Error('An unknown error occurred during conversion');
 	}
 }
